sistemato problema della compilazione in 64 bit che dava risultati sbagliati. Bastava lasciare size_t invece di unsigned long int e il rispettivo SIZE_MAX invece di ULONG_MAX.

+++ BeliefPropagationCUDA/myLibKernelSmallN.h
@@ -1,4 +1,6 @@
-double * runSmallN(int size, int nArray, double *h_idata, long unsigned int *h_iIndexData);
+
+
+double * runSmallN(int size, int nArray, double *h_idata, size_t *h_iIndexData);
 
 void selectGPU();
 void resetGPU();
